requirements.txt
asgiref==3.3.1
astroid==2.5.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
configparser==5.0.2
Django==3.1.7
django-appconf==1.0.4
django-ckeditor==6.0.0
django-debug-toolbar==3.2
django-imagekit==4.0.2
django-js-asset==1.2.2
idna==2.10
isort==5.7.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
pilkit==2.0
Pillow==8.1.2
pylint==2.7.2
pylint-django==2.4.2
pylint-plugin-utils==0.6
PyMySQL==1.0.2
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.3
wrapt==1.12.1
